Fixed Android build missing its dependencies
https://github.com/googlesamples/unity-jar-resolver/issues/384
The cause was the lack of required dependencies. The Android Manager's Force Resolver would throw a ClassNotFoundException when starting a Gradle daemon. Fixed it by removing another installation of Java SDK used in another project, because apparently Unity was trying to use that SDK, instead of OpenJDK it has downloaded.